feat: Update CustomIconButton component to use custom icon color
The code changes in `CustomIconButton.tsx` modify the `CustomIconButton` component to allow for a custom icon color. The `iconColor` prop has been updated to accept a hex color value (`#fff`) in addition to the predefined color options. This change provides more flexibility in styling the icon color.

The commit message suggests updating the `CustomIconButton` component to use a custom icon color.