refactor: Add dashboardType constant for Table and Graph options
The code changes in constants.ts add a new constant `dashboardType` with options for "Table" and "Graph". This constant will be used to determine the type of dashboard to display.

This commit message follows the established convention of starting with a verb in the imperative form ("Add dashboardType constant for Table and Graph options") and provides a clear and concise summary of the changes made.