refactor: Update InvoiceDetail component with constants and API hook
This commit updates the InvoiceDetail component by adding the necessary imports for the Constants and useInvoicesettingsControllerFindFirst. The Constants are used to retrieve the invoiceDefaultPrefix, which is then displayed in the Typography component. The useInvoicesettingsControllerFindFirst hook is imported for future use in fetching invoice settings data. This change improves the readability and maintainability of the code.

Co-authored-by: chandrajyoti123 <chandrajyotiadil15062003@gmail.com>