feat: Update useNotificationSocket to connect to secure server with reconnection options
The useNotificationSocket hook has been updated to connect to a secure server using the "https://growinvoice-94ee0dd2031b.herokuapp.com" URL. The socket connection now includes options for reconnection, with 5 reconnection attempts and a delay of 1000 milliseconds between each attempt. This change improves the reliability and stability of the socket connection.

Co-authored-by: chandrajyoti123 <chandrajyotiadil15062003@gmail.com>