refactor: Remove GetStartedDialog from AppContainer
The GetStartedDialog component has been removed from the AppContainer in order to improve the user experience. This dialog was previously shown when the user's company information was incomplete, but it has been decided to remove this feature. The dialog is no longer necessary and has been removed from the codebase.